working backend with frontend

+++ backend/src/main/java/hu/unideb/movilist/data/repository/StatusRepository.java
@@ -1,3 +1,4 @@
+
 package hu.unideb.movilist.data.repository;
 
 import hu.unideb.movilist.data.entity.Status;
@@ -5,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StatusRepository extends JpaRepository<Status, Long> {
-}
+public interface StatusRepository extends JpaRepository<Status, Integer> {}
+
